detect lutris intall

+++ waven-linux.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 dir="$PWD"
+lutriswinepath="/home/$USER/.local/share/lutris/runners/wine"
 
 # change version number to downgrade to another old build
 winever="6.14-4"
@@ -36,25 +37,41 @@ dxvk)
 esac
 
 configure() {
+  if [ -d $lutriswinepath ]; then
+    if [ -d "$lutriswinepath/lutris-$winever-x86_64" ]; then
+      if [ ! -d "$dir/lutris-$winever-x86_64" ]; then
+          ln -s $lutriswinepath/lutris-$winever-x86_64 $dir
+          wineinstall=true
+      fi
+    echo "Wine correctement installé"
+    else
+      echo "Télécharger wine depuis Lutris, version : lutris-$winever"
+      echo "Puis relancer le script"
+    fi
+  else
     # download lutris wine build
-    if [ ! -d "lutris-$winever-x86_64" ]; then
+    if [ ! -d "$winever-x86_64" ]; then
         wget https://github.com/lutris/wine/releases/download/lutris-$winever/wine-lutris-$winever-x86_64.tar.xz
         tar -xf wine-lutris-$winever-x86_64.tar.xz
         rm wine-lutris-$winever-x86_64.tar.xz
+        wineinstall=true
     fi
+  fi
 
+  if [ "$wineinstall" = true ]; then
     # configure wine prefix, install launcher
     if [ ! -d "wine-waven" ]; then
-        wget "https://launcher.cdn.ankama.com/installers/production/Ankama Launcher-Setup.exe" -O "ankama-launcher.exe"
+      wget "https://launcher.cdn.ankama.com/installers/production/Ankama Launcher-Setup.exe" -O "ankama-launcher.exe"
 
-    WINEPREFIX=$prefix WINEDLLOVERRIDES=$overrides "$wine" "$exec"
-    rm ankama-launcher.exe
+      WINEPREFIX=$prefix WINEDLLOVERRIDES=$overrides "$wine" "$exec"
+      rm ankama-launcher.exe
     fi
 
     echo "$script" | tee start-waven.sh
 
     # add execute to script
     chmod +x $dir/start-waven.sh
+  fi
 }
 
 # configure dxvk
